Move structures to their own files
